refactor: standardizes imports and corrects the order of dependencies in the adapter module
